moving the `reflection` module to a shared space
