+++ src/core/reflection/__test__/filter-scope.spec.ts
