+++ src/core/reflection/inheritance-chain.ts
